% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dsb_courses.R
\name{install_dsb_courses}
\alias{install_dsb_courses}
\title{Install (German) DSB swirl courses}
\usage{
install_dsb_courses(courses = "all", force = FALSE)
}
\arguments{
\item{courses}{character; single string or vector with the name of the courses to be
installed. The default setting is 'all', which installs all German courses, but you
can also select from the following courses:
'DSB-01', 'DSB-02', 'DSB-03', 'DSB-04', 'DSB-05', 'DSB-06', 'Data analysis with R'.
See \emph{Details} for the topics each course covers.}

\item{force}{logical; should course installation be forced if the course is already
installed? The default value is \code{FALSE}.}
}
\description{
Install all or a selection of swirl courses developed for the DSB (Data Science in
Biology) program within the biology department of the University of Hamburg (UHH).
The package currently includes seven swirl courses that can be installed
using this \code{install_dsb_courses} function.
}
\details{
The following swirl courses can be installed (currently only in German):
\describe{
  \item{DSB-01}{R Grundlagen (Basics in R)}
  \item{DSB-02}{Datenexploration mit R (Data exploration in R))}
  \item{DSB-03}{Datenaufbereitung oder per Anleitung durchs Tidyversum
                (Data wrangling and introduction to tidyverse)}
  \item{DSB-04}{Datenvisualisierung mit ggplot2 (Data visualization with ggplot2)}
  \item{DSB-05}{Handling spezieller Datentypen (Working with special data types)}
  \item{DSB-06}{Fortgeschrittene R Programmierung (Advanced R programming)}
}
The following swirl course is in English:
\describe{
  \item{Data analysis with R}{From R basics to visualization and statistical modelling}
 }
}
\examples{
\dontrun{
 # Install all German courses (default)
 install_dsb_courses()

 # Install only the DSB-01 (R basics) course
 install_dsb_courses(courses = "DSB-01")

 # Install the courses DSB-01, DSB-04, and DSB-05
 install_dsb_courses(courses = c("DSB-01", "DSB-04", "DSB-05"))
}
}
