% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dsb_courses.R
\name{install_dsb_courses}
\alias{install_dsb_courses}
\title{Install (German) DSB swirl courses}
\usage{
install_dsb_courses(courses = "all", force = FALSE)
}
\arguments{
\item{courses}{character; single string or vector with the name of the courses to be
installed. By default, all courses will be installed but you can also select from the
the following courses: 'DSB-01', 'DSB-02', 'DSB-03', 'DSB-04', 'DSB-05'. See under
details what topics each course covers.}

\item{force}{logical; should course installation be forced if the course is already
installed? The default value is FALSE.}
}
\description{
Install all or a selection of swirl courses developed for the DSB (Data Science in
Biology) program within the biology department of the University of Hamburg (UHH).
The package comes currently with five swirl courses that can be installed
using this \code{install_dsb_courses} function.
}
\details{
The following swirl courses can be installed (currently only in German):
\describe{
  \item{DSB-01}{Basics in R (R Grundlagen)}
  \item{DSB-02}{Data exploration with R (Datenexploration mit R)}
  \item{DSB-03}{Data warangling and introduction to tidyverse (Datenaufbereitung
    oder per Anleitung durchs Tidyversum)}
  \item{DSB-04}{Data visualization with ggplot2 (Datenvisualisierung mit ggplot2)}
  \item{DSB-05}{Handling of special data types (Handling spezieller Datentypen)}
 }
}
\examples{
\dontrun{
 # Install all courses (default)
 install_dsb_courses()
 # Install only course DSB-01 (R basics)
 install_dsb_courses(courses = "DSB-01")
 # Install the course DSB-01, DSB-04, and DSB-05
 install_dsb_courses(courses = c("DSB-01", "DSB-04", "DSB-05"))
}
}
